Move the Application model into Providers module
The `Models` module is eventually going to be merged with the providers
as a way to remove the coupling between the two.

The first step in doing so is to move the `Application` class into the
`Providers::Base` module
